requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
deep-translator>=1.11.4
feedparser>=6.0.0
pillow>=10.0.0
langdetect>=1.0.9
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
werkzeug>=3.0.0
flask-limiter>=3.5.0
flask-caching>=2.1.0
bcrypt>=4.1.0
certifi>=2024.0.0
